third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/scroll-state/scroll-state-scroll-direction-wm.html

<!DOCTYPE html>
<title>@container: scroll-state(scroll-state) matching for writing direction</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/6400">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<script src="/css/css-transitions/support/helper.js"></script>
<style>
  .scroll-direction {
    container-type: scroll-state;
    width: 100px;
    height: 100px;
    overflow: scroll;
  }
  .filler {
    height: 600px;
    width: 600px;
  }
  .target {
    @container scroll-state(scroll-direction: block-start) { --block-start: yes }
    @container scroll-state(scroll-direction: block-end) { --block-end: yes }
    @container scroll-state(scroll-direction: inline-start) { --inline-start: yes }
    @container scroll-state(scroll-direction: inline-end) { --inline-end: yes }
    @container scroll-state(scroll-direction: top) { --top: yes }
    @container scroll-state(scroll-direction: left) { --left: yes }
    @container scroll-state(scroll-direction: bottom) { --bottom: yes }
    @container scroll-state(scroll-direction: right) { --right: yes }
  }
  .ltr { direction: ltr; }
  .rtl { direction: rtl; }
  .htb { writing-mode: horizontal-tb; }
  .vlr { writing-mode: vertical-lr; }
  .vrl { writing-mode: vertical-rl; }
</style>
<div class="scroll-direction h htb ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction h htb rtl"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction h vlr ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction h vlr rtl"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction h vrl ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction h vrl rtl"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v htb ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v htb rtl"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v vlr ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v vlr rtl"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v vrl ltr"><div class="filler"><span class="target"></span></div></div>
<div class="scroll-direction v vrl rtl"><div class="filler"><span class="target"></span></div></div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  function match_scroll_direction(container_selector, expected_matches) {
    let scroller = document.querySelector(container_selector + " .target");
    let style = getComputedStyle(scroller);
    for (let custom_prop of [ "--block-start",
                              "--block-end",
                              "--inline-start",
                              "--inline-end",
                              "--top",
                              "--left",
                              "--bottom",
                              "--right" ]) {
      assert_equals(style.getPropertyValue(custom_prop) === "yes",
                    expected_matches[custom_prop] === true, custom_prop);
    }
  }

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.htb.ltr");
    scroller.scrollLeft = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.h.htb.ltr", { "--inline-end": true, "--right": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar horizontal-tb/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.htb.rtl");
    scroller.scrollLeft = -150;
    await waitForAnimationFrames(2);
    assert_equals(scroller.scrollLeft, -150);
    match_scroll_direction(".scroll-direction.h.htb.rtl", { "--inline-end": true, "--left": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar horizontal-tb/rtl");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.vlr.ltr");
    scroller.scrollLeft = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.h.vlr.ltr", { "--block-end": true, "--right": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar vertical-lr/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.vlr.rtl");
    scroller.scrollLeft = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.h.vlr.rtl", { "--block-end": true, "--right": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar vertical-lr/rtl");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.vrl.ltr");
    scroller.scrollLeft = -150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.h.vrl.ltr", { "--block-end": true, "--left": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar vertical-rl/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.h.vrl.rtl");
    scroller.scrollLeft = -150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.h.vrl.rtl", { "--block-end": true, "--left": true });
  }, "scroll-state(scroll-direction) horizontal scrollbar vertical-rl/rtl");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.htb.ltr");
    scroller.scrollTop = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.htb.ltr", { "--block-end": true, "--bottom": true });
  }, "scroll-state(scroll-direction) vertical scrollbar horizontal-tb/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.htb.rtl");
    scroller.scrollTop = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.htb.rtl", { "--block-end": true, "--bottom": true });
  }, "scroll-state(scroll-direction) vertical scrollbar horizontal-tb/rtl");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.vlr.ltr");
    scroller.scrollTop = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.vlr.ltr", { "--inline-end": true, "--bottom": true });
  }, "scroll-state(scroll-direction) vertical scrollbar vertical-lr/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.vlr.rtl");
    scroller.scrollTop = -150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.vlr.rtl", { "--inline-end": true, "--top": true });
  }, "scroll-state(scroll-direction) vertical scrollbar vertical-lr/rtl");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.vrl.ltr");
    scroller.scrollTop = 150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.vrl.ltr", { "--inline-end": true, "--bottom": true });
  }, "scroll-state(scroll-direction) vertical scrollbar vertical-rl/ltr");

  promise_test(async t => {
    let scroller = document.querySelector(".scroll-direction.v.vrl.rtl");
    scroller.scrollTop = -150;
    await waitForAnimationFrames(2);
    match_scroll_direction(".scroll-direction.v.vrl.rtl", { "--inline-end": true, "--top": true });
  }, "scroll-state(scroll-direction) vertical scrollbar vertical-rl/rtl");

</script>
